Create CircleMouseFollow.java
New assignment. The ball has to follow the mouse but slow down once it gets close to it. The ball will also turn red if the mouse touches the ball. There will be another file with this as the second part to the assignment is to make the ball move constantly.